Run loops in foreground to use same namespace.
We're using a named pipe to communicate with mpb(1) without
messing with stdout. Previously, mpb(1) was running in the
foreground, interacting with the loop which was in background.
However, the loop needs to update variables from the script
that will be used after it finishes.
This change fixes the count of removed/(un)masked/modified files.
Turns out it's also simpler than the previous approach.